requirements.txt
neo4j>=5.14.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
google-genai
tenacity>=8.2.0
SPARQLWrapper>=2.0.0
thefuzz>=0.20.0
python-Levenshtein>=0.21.0
